+++ examples/02-selection.cpp
@@ -10,9 +10,9 @@ typedef ea::Sequence<int32_t> Sequence;
 class Fitness
 {
 	public:
-		float operator()(const Sequence* const &seq) const
+		double operator()(const Sequence* const &seq) const
 		{
-			float avg = 0;
+			double avg = 0;
 
 			for(ea::sequence_len_t i = 0; i < ea::sequence_len(seq); ++i)
 			{
